Migrated backpack consumers to logical storage

+++ src/main/java/io/github/tootertutor/ModularPacks/gui/AdminBackpackListMenuHolder.java
@@ -136,7 +136,7 @@ public final class AdminBackpackListMenuHolder implements InventoryHolder {
             String backpackName,
             String ownerUuid,
             String ownerName,
-            int itemCount,
+            long itemCount,
             int moduleCount,
             boolean placed,
             String locationText,
